+++ inst/shinybrms_app/tests/shinytest/switchData_same.R
@@ -1,0 +1,33 @@
+app <- ShinyDriver$new("../../")
+app$snapshotInit("switchData_same", screenshot = FALSE)
+
+app$setInputs(advOpts_cores = 4, wait_ = FALSE, values_ = FALSE)
+
+app$setInputs(navbar_ID = "Data")
+app$uploadFile(file_upload = "switchData-data.csv")
+app$setInputs(navbar_ID = "Likelihood",
+              outc_sel = "y",
+              dist_sel = "bernoulli")
+app$setInputs(likelihood_navlist_ID = "Predictors",
+              pred_mainNP_sel = "x1")
+app$setInputs(navbar_ID = "Prior",
+              prior_class_sel = "b")
+app$setInputs(prior_text = "normal(0, 2.5)",
+              prior_add = "click")
+app$snapshot(items = list(input = setdiff(app$listWidgets()$input, "file_upload"),
+                          output = TRUE,
+                          export = TRUE))
+app$setInputs(navbar_ID = "Data")
+app$uploadFile(file_upload = "switchData-data.csv")
+app$snapshot(items = list(input = setdiff(app$listWidgets()$input, "file_upload"),
+                          output = TRUE,
+                          export = TRUE))
+app$setInputs(navbar_ID = "Prior")
+app$snapshot(items = list(input = setdiff(app$listWidgets()$input, "file_upload"),
+                          output = TRUE,
+                          export = TRUE))
+app$setInputs(navbar_ID = "Likelihood")
+app$setInputs(likelihood_navlist_ID = "Outcome")
+app$snapshot(items = list(input = setdiff(app$listWidgets()$input, "file_upload"),
+                          output = TRUE,
+                          export = TRUE))